+++ src/Module/System/Update.php
@@ -825,6 +825,9 @@ class Update
         $update_string = "* Add `object_type_IDX` to artist_map table<br />* Add `object_type_IDX` to catalog_map table";
         $version[]     = array('version' => '600016', 'description' => $update_string);
 
+        $update_string = "* Drop `user_playlist` table and recreate it";
+        $version[]     = array('version' => '600017', 'description' => $update_string);
+
         return $version;
     }
 
@@ -5068,6 +5071,25 @@ class Update
         $sql = "ALTER TABLE `catalog_map` DROP KEY `object_type_IDX`;";
         Dba::write($sql);
         $sql = "CREATE INDEX `object_type_IDX` USING BTREE ON `catalog_map` (`object_type`);";
+        $retval &= (Dba::write($sql) !== false);
+
+        return $retval;
+    }
+
+    /**
+     * update_600017
+     *
+     * Drop `user_playlist` table and recreate it
+     */
+    public static function update_600017(): bool
+    {
+        $retval    = true;
+        $collation = (AmpConfig::get('database_collation', 'utf8mb4_unicode_ci'));
+        $charset   = (AmpConfig::get('database_charset', 'utf8mb4'));
+        $engine    = ($charset == 'utf8mb4') ? 'InnoDB' : 'MYISAM';
+        $sql       = "DROP TABLE IF EXISTS `user_playlist`";
+        Dba::write($sql);
+        $sql = "CREATE TABLE IF NOT EXISTS `user_playlist` (`playqueue_time` int(11) UNSIGNED NOT NULL, `playqueue_client` varchar(255) CHARACTER SET $charset COLLATE $collation DEFAULT NULL, `user` int(11) DEFAULT NULL, `object_type` enum('song', 'live_stream', 'video', 'podcast_episode')  CHARACTER SET utf8 COLLATE utf8_unicode_ci DEFAULT NULL, `object_id` int(11) UNSIGNED NOT NULL DEFAULT 0, `track` smallint(6) DEFAULT NULL, `current_track` tinyint(1) UNSIGNED NOT NULL DEFAULT 0, `current_time` smallint(5) UNSIGNED NOT NULL DEFAULT 0, PRIMARY KEY (`playqueue_time`,`playqueue_client`,`user`,`track`), KEY `user` (`user`), KEY `object_type` (`object_type`), KEY `object_id` (`object_id`)) ENGINE=$engine DEFAULT CHARSET=$charset COLLATE=$collation;";
         $retval &= (Dba::write($sql) !== false);
 
         return $retval;